+++ CMakeLists.txt
@@ -164,6 +164,7 @@ include_directories("/usr/local/include")
 link_directories("/usr/local/lib")
 
 add_subdirectory(osquery)
+add_subdirectory(tools/tests)
 
 # make docs
 find_package(Doxygen)
