+++ osquery/CMakeLists.txt
@@ -55,7 +55,6 @@ add_subdirectory(events)
 add_subdirectory(extensions)
 add_subdirectory(filesystem)
 add_subdirectory(logger)
-add_subdirectory(python_tests)
 add_subdirectory(registry)
 add_subdirectory(scheduler)
 add_subdirectory(sql)
