--- osquery/python_tests/test_base.py
@@ -1,59 +1,0 @@
-import re
-
-import pexpect
-from pexpect.replwrap import REPLWrapper
-
-
-class OsqueryUnknownException(Exception):
-    '''Exception thrown for unknown output from the shell'''
-    pass
-
-
-class OsqueryException(Exception):
-    '''Exception thrown when the shell returns an error'''
-    pass
-
-
-class OsqueryWrapper(REPLWrapper):
-    '''A pexpect wrapper intended for interacting with the osqueryi REPL'''
-    PROMPT = u'osquery> '
-    CONTINUATION_PROMPT = u'    ...> '
-    ERROR_PREFIX = 'Error:'
-
-    def __init__(self, command='../osqueryi'):
-        super(OsqueryWrapper, self).__init__(
-            command,
-            self.PROMPT,
-            None,
-            continuation_prompt=self.CONTINUATION_PROMPT)
-
-    def run_query(self, query):
-        '''Run a query, returning the results as a list of dictionaries
-
-        When unknown output is encountered, OsqueryUnknownException is thrown.
-        When osqueryi returns an error, OsqueryException is thrown.
-        '''
-        query = query + ';'  # Extra semicolon causes no harm
-        result = self.run_command(query)
-        # On Mac, the query appears first in the string. Remove it if so.
-        result = re.sub(re.escape(query), '', result).strip()
-        result_lines = result.splitlines()
-
-        if len(result_lines) < 1:
-            raise OsqueryUnknownException(
-                'Unexpected output:\n %s' % result_lines)
-        if result_lines[0].startswith(self.ERROR_PREFIX):
-            raise OsqueryException(result_lines[0])
-
-        try:
-            header = result_lines[1]
-            columns = re.findall('[^ |]+', header)
-            rows = []
-            for line in result_lines[3:-1]:
-                values = re.findall('[^ |]+', line)
-                rows.append(
-                    dict((col, val) for col, val in zip(columns, values)))
-            return rows
-        except:
-            raise OsqueryUnknownException(
-                'Unexpected output:\n %s' % result_lines)
