+++ tools/tests/CMakeLists.txt
@@ -1,0 +1,4 @@
+# List of python modules containing unittests
+ADD_OSQUERY_PYTHON_TEST(test_osqueryi test_osqueryi.py)
+ADD_OSQUERY_PYTHON_TEST(test_watchdog test_watchdog.py)
+ADD_OSQUERY_PYTHON_TEST(test_extensions test_extensions.py)
